Fix pandas/PETSc issues (#1136)
* Fix tests

* Remove `is_categorical_dtype`

* Fix type checking

* Fix converged flag in `PETSc>=3.20`

* Don't pass dtype to `AnnData`

* Start fixing warnings

* Fix `pandas.__getitem__` warnings

* Correct serie's dtype

* Fix map -> apply

* No duplicate names

* Fix `DataFrame.applymap`

+++ src/cellrank/kernels/_precomputed_kernel.py
@@ -142,7 +142,7 @@ class PrecomputedKernel(UnidirectionalKernel):
         # fmt: off
         if adata is None:
             logg.warning(f"Creating empty `AnnData` object of shape `{matrix.shape[0], 1}`")
-            adata = AnnData(sp.csr_matrix((matrix.shape[0], 1), dtype=np.float64))
+            adata = AnnData(sp.csr_matrix((matrix.shape[0], 1)))
         super().__init__(adata)
         self._backward: Optional[bool] = backward
         self.transition_matrix = matrix.copy() if copy else matrix
